deconflicting and updating unverified commits
