+++ .gitignore
@@ -26,5 +26,14 @@ Thumbs.db
 # Optional: Exclude requirements.txt if using pyproject.toml exclusively
 requirements.txt
 
+<<<<<<< HEAD
 # environment variables
 .env
+=======
+# exclude internal docs
+refactoring-guidance.md
+cluster-snek_proj_asmt.pdf
+
+# exclude scratchpad dir
+scratchpad-wip
+>>>>>>> 2bd2747 (Convergence (#3))
